requirements.txt
amqp==5.0.1
asgiref==3.3.0
billiard==3.6.3.0
celery==5.0.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
docker==4.3.1
idna==2.10
importlib-metadata==2.0.0
kombu==5.0.2
more-itertools==8.5.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
urllib3==1.25.11
vine==5.0.0
websocket-client==0.56.0
zipp==3.3.1
